solution hope final 😁
