+++ src/index.tsx
@@ -4,5 +4,10 @@ import 'bulma/css/bulma.css';
 import '@fortawesome/fontawesome-free/css/all.css';
 import './styles/index.scss';
 import { App } from './App';
+import { TodoContext } from './TodoContext';
 
-createRoot(document.getElementById('root') as HTMLDivElement).render(<App />);
+createRoot(document.getElementById('root') as HTMLDivElement).render(
+  <TodoContext>
+    <App />
+  </TodoContext>,
+);
